refactor(httpapi): #220 move request handlers out of server.go
Pure move of server.go:55-496 (handleHealth through apiHandler) into
handlers.go; server.go keeps Server assembly, routing, and parse/JSON
helpers. Byte-identity verified against HEAD. No behavior change.
